fix: split navigation main file in two
